+++ rails_32/spec/support/database_cleaner.rb
@@ -1,3 +1,6 @@
+# Good read on using database_cleaner (in lieu of a shared connection):
+#   http://devblog.avdi.org/2012/08/31/configuring-database_cleaner-with-rails-rspec-capybara-and-selenium/
+
 RSpec.configure do |config|
   config.before(:suite) do
     DatabaseCleaner.clean_with(:truncation)