+++ rails_40/app/controllers/pages_controller.rb
@@ -1,0 +1,6 @@
+class PagesController < ApplicationController
+
+  def root
+  end
+
+end