+++ rails_40/config/routes.rb
@@ -1,0 +1,5 @@
+AppPrototype::Application.routes.draw do
+
+  root to: 'pages#root'
+
+end